langeval-core/identity-service/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-jose[cryptography]>=3.3.0
httpx>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
sqlmodel>=0.0.8
passlib[bcrypt]>=1.7.4
google-auth>=2.0.0
requests>=2.0.0